contracts: fetch contracts data on server
